Split tail components into two
